Move test files to proper dir
